requirements.txt
langchain>=0.1.16
langchain-community>=0.0.32
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.3

faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
streamlit>=1.30.0
